Ifindex support in iris (#19867)
* Ifindex support in iris

* fixing port ifindex in agent

* Fixing port issues

* Fixing gtests and dol

* Fixing ethtool

* Fixing dols

* fixing gft

* fixing go

* commenting out e2e naples sim

* fixing esx

+++ nic/fte/test/fte_base_test.cc
@@ -149,7 +149,7 @@ hal_handle_t fte_base_test::add_enic(hal_handle_t l2segh, uint32_t useg, uint64_
     return resp.mutable_status()->if_handle();
 }
 
-hal_handle_t fte_base_test::add_uplink(uint8_t port_num)
+hal_handle_t fte_base_test::add_uplink(uint32_t port_num)
 {
     hal_ret_t ret;
     intf::InterfaceSpec spec;
